1.create 'ByteStringExtensions' 2.rename ComputeHash to ToHash 3.move 'DumpByteArray' to 'NumbericExtensions'
