+++ contract/AElf.Contracts.Genesis/BasicContractZero.cs
@@ -211,7 +211,7 @@ namespace AElf.Contracts.Genesis
 
         public static Address BuildContractAddress(int chainId, ulong serialNumber)
         {
-            return BuildContractAddress(chainId.ComputeHash(), serialNumber);
+            return BuildContractAddress(chainId.ToHash(), serialNumber);
         }
     }
 }
