+++ src/AElf.Kernel.Core/Blockchain/Infrastructure/IStaticChainInformationProvider.cs
@@ -42,7 +42,7 @@ namespace AElf.Kernel.Blockchain.Infrastructure
 
         public static Address BuildContractAddress(int chainId, ulong serialNumber)
         {
-            return BuildContractAddress(chainId.ComputeHash(), serialNumber);
+            return BuildContractAddress(chainId.ToHash(), serialNumber);
         }
     }
 }
