+++ src/AElf.Types/Extensions/HashExtensions.cs
@@ -5,7 +5,7 @@ namespace AElf
 {
     public static class HashExtensions
     {
-        public static Hash ComputeHash(this int obj)
+        public static Hash ToHash(this int obj)
         {
             return Hash.FromRawBytes(BitConverter.GetBytes(obj));
         }
