+++ src/AElf.Types/Extensions/HexExtensions.cs
@@ -4,9 +4,6 @@ namespace AElf
 {
     public static class HexExtensions
     {
-        public static byte[] DumpByteArray(this int n)
-        {
-            return BitConverter.GetBytes(n);
-        }
+
     }
 }
