+++ src/AElf.Types/Extensions/NumbericExtensions.cs
@@ -9,5 +9,9 @@ namespace AElf
         {
             return BitConverter.IsLittleEndian ? BitConverter.GetBytes(number).Reverse().ToArray() : BitConverter.GetBytes(number);
         }
+        public static byte[] DumpByteArray(this int n)
+        {
+            return BitConverter.GetBytes(n);
+        }
     }
 }
